Bug 1709269 - [devtools] Make type optional in TargetCommand::hasTargetWatcherSupport r=nchevobbe
The rest of the patches in this stack will remove WatcherActor traits.
But the call sites should still check that the watcher is available.
Making the argument optional allows to reuse the same method as we remove the traits.

Differential Revision: https://phabricator.services.mozilla.com/D114216

UltraBlame original commit: dd1897d8f4f7b677941960a55e862efca4584c74