Bug 1513447 - Use generated marker elements for list bullet accessibles. r=Jamie
In 1513447 there is a demonstrated instance in which the generated
marker is replaced with another one and throws the list item bullet
state into an unknown state. To remedy this we need to observe when such
elements are removed and added.

Instead of that, I opted to finally make the bullet accessible a real
content-backed accessible. This should help with other issues that pop
up when the list item overrides children management and keeps an
artificial accessible as its first child.

Differential Revision: https://phabricator.services.mozilla.com/D110719

UltraBlame original commit: 5ff1cc0ecdb3cc503d6e0b6f5509eda78d551c7a