Bug 1708985 - enforce the default background colour on the ContentSelect menulist's menupopup when custom option foreground colours are used, r=emilio,jaws
Broadly, this patch does 2 things:

- it mirrors the logic bug 1702258 introduced for <select> styling for <option> styling
- it enforces that if the website specifies custom foreground colours for options,
  we use the website-specified colour for the option background, too (even if it
  matches the UA/select style).

I also added automated testing for 3 cases:
1. the default dark mode situation (where we expect dark mode styles),
2. the case from the bug where the website specifies `#fff` for the background
of the select and the background of the options, but specifies a custom
foreground colour for the option (but not the select).
3. a very similar case where the website only specifies the background on the
select, and only the foreground on the option.

Differential Revision: https://phabricator.services.mozilla.com/D114382

UltraBlame original commit: 75badaa5291a4d5a4b53b81e8d6b4750cc845e79