Bug 1709606 - Move GfxInfoX11 into gtk and remove X11 from the name, r=stransky
It is only used by the GTK backend, supports Wayland as well and
is highly unlikely to be reused by another backend.

This is also better in line with how this is handled for other
platforms.

While on it, do some small cleanups.

Depends on D114423

Differential Revision: https://phabricator.services.mozilla.com/D114424

UltraBlame original commit: 059be8a083a3bc1da86ee7a285fb7bbf4678f708