Bug 1704981 - [devtools] initialize the data provider on target available r=nchevobbe,ochameau
After some investigations,I'l try to explain what i see. What seems to be happening is that the data provider is
intialized when the webconsole-ui is initialized  so that is when the instance of the console
front is created. When a target switch happens, that instance of the console front gets destroyed
so when the data provider needs to request the network details it fails and this._client
is null.

This patch moves the creation of the data provider to onTargetAvailable (as we do with that in the network panel).
This fixes the issue, and TRY seems to be good with it, I wonder if this is best solution or this can be improved.
What are your thoughts?
I can add a test if we think this is the way to go.

Differential Revision: https://phabricator.services.mozilla.com/D113167

UltraBlame original commit: 2e702b9402b87e6b48fbe65e4d992a24e3bcfd10