Bug 1707022 - p3: Focus open or close buttons when menu is toggled. r=amy
The `timeout` for the CSSTransition needs to match the actual duration
of the CSS transition. This is needed because the close button needs to
be visible before focus() is called on it or else it won't recieve
focus.

Differential Revision: https://phabricator.services.mozilla.com/D113867

UltraBlame original commit: a223e40d5d1626b8e5f6a16c490a87cb9d0821ac