Bug 1707022 - p5: Clean up menu transition CSS. r=amy
I took the liberty to clean up the CSS a bit. There is no need to have
special exceptions for how `visibility` is animated. Since this is a
binary property it becomes invisible at the end of a transition
duration, and visible at the begining.

Differential Revision: https://phabricator.services.mozilla.com/D113869

UltraBlame original commit: 285861b3293420abaad9fa272b060f1fe15fcf7e