Bug 1709415 - Support -webkit-image-set as a parse-time alias to image-set(). r=karlcow,twisniewski
The webkit syntax is an strict subset of the modern one, so this should
be doable, and is the simplest.

If my reading of the WebKit code is correct it should also be the way
WebKit deals with this (except they restrict -webkit-image-set() syntax
artificially).

 * https://github.com/w3c/csswg-drafts/issues/6285
 * https://github.com/whatwg/compat/issues/144

Differential Revision: https://phabricator.services.mozilla.com/D114912

UltraBlame original commit: 01bca848cb09b426c0c1c5f34d32dbe3fa5b9aa7