Bug 1710607 - If not profiling, make ShutdownProfilerChild dispatch SYNC - r=canaltinova
When not profiling, ShutdownProfilerChild should be quick, so it's safe to run synchronously.
In this case, we don't need to manually shutdown the thread (to wait for the previously-async task).
This removes the risky thread-shutdown with its inner event-handling loop, where all crashes in bug 1613798 happen.

Differential Revision: https://phabricator.services.mozilla.com/D114845

UltraBlame original commit: 02ea24efbd7c484246b2a9aa161c0b8bb1cb56e6