Bug 1711164 - Ensure we never use cached surfaces with blob recordings. r=gfx-reviewers,nical
We don't support storing blob recordings in the surface cache at this
time so we should never look at it. It is possible that one may request
a blob recording, and have a previous rasterization available in the
cache because it was used in canvas or as a repeating background image.
We want to explicitly avoid the cache in that case.

Differential Revision: https://phabricator.services.mozilla.com/D115136

UltraBlame original commit: 7811b10f32a1f0ac9c0b1044e14850b2757ded3a