Bug 1709591 - Use arrow-dropdown-16.svg for the Synced Tabs expander icon rather than rotating a right-pointing one. r=mhowell
Before this patch, we were showing a downward-pointing expander icon by taking a
rightward-pointing one and rotating it 90deg. That's fine, except for RTL, where we
tend to mirror rightward-pointing arrows with a `scale: -1 1;` rule. Because of
order of operations, the scale rule is applied first, and the 90deg rotation
that occurs afterwards ends up pointing the scaled arrow upwards.

Instead of special-casing the scale rule for this expander icon, I've gone ahead
and switched it to use arrow-dropdown-16.svg and gotten rid of the rotation. This
way, even if we mirror across the y-axis with `scale: -1 1`, the appearance will
not change.

Differential Revision: https://phabricator.services.mozilla.com/D115268

UltraBlame original commit: e69e2aecb3dce03d3adb58cc64b4c76e8d23454f