Bug 1627175 - part 62: Get rid of `HTMLEditor::IsPrevCharInNodeWhiteSpace()` r=m_kato
It's used only by `HTMLEditor::GetWhiteSpaceEndPoint()` and it just skips
white-spaces and returns first white-space position if and only if the
given point container is a text node and its previous character is a
white-space.  Therefore, it can be rewritten with `EditorDOMPoint` simply in
the caller.

Differential Revision: https://phabricator.services.mozilla.com/D115171

UltraBlame original commit: 6a05527852751ddd4ce377226fe5968663c90bc7