Bug 1627175 - part 63: Get rid of `HTMLEditor::IsNextCharInNodeWhiteSpace()` r=m_kato
Similarly, it's called only by `HTMLEditor::GetWhiteSpaceEndPoint()` and
it returns after point of last white-space if and only if the given point
container is a text node and the offset is not end of the text node.
Therefore, we can reimplement it in the caller simply.

Differential Revision: https://phabricator.services.mozilla.com/D115172

UltraBlame original commit: ea32cc59ac32699498e5cd6546a04063ded062ea