Bug 1711211 - Ensure RenderCompositorOGLSWGL does a full render after EGL surface creation. r=sotaro
RenderCompositorOGLSWGL will currently skip compositing if webrender
decides that the dirty rect for a frame is empty. However, in cases
where the EGL surface has been recreated we must perform a full
composite, as the previous contents of the framebuffer has been
lost. Not doing so is resulting in black screens being shown after
restoring the app on android, or the contents of the previous tab
being shown after switching to a new tab.

In the future we will add proper partial rendering support by tracking
buffer age and damage regions, and webrender will be able to force a
full render when the buffer age is zero. Until that time, we can fix
this bug by making RenderCompositorOGLSWGL::RequestFullRender() return
true for the subsequent frame after the surface has been recreated.

Differential Revision: https://phabricator.services.mozilla.com/D115342

UltraBlame original commit: 1273a2970c98aac86e07707ca487c13fda6545bf