Bug 1711636 - Assertion failure: from.toStackSlot()->slot() % SimdMemoryAlignment == 0, at js/src/jit/LIR.cpp:654.  r=lth.
This assertion is overly constraining: in fact we don't require or provide
naturally-aligned 128-bit accesses to the stack on arm64.  So disable it for
arm64, and cause a compilation failure on any target supporting wasm SIMD that
isn't x86, x64 or arm64.  The original alignment-check code remains but is
commented out, just in case we need to reinstate it at some point.

Differential Revision: https://phabricator.services.mozilla.com/D115466

UltraBlame original commit: 69ebfd62d6a364f97903d3bf9ab05896aec6f770