Bug 1709851 - Update fallback GMP downloader for Widevine 4.10.2209.1. r=mjf
Update the GMP fallback downloader for Widevine to reference the 4.10.2209.1
CDM. This keeps the fallback URLs in step with Balrog.

Add the aarch64/arm64 entry for the new Apple Silicone CDM to the fallback.

Reorder the json file so it matches the ordering used by balrog so it's easier
to eyeball and diff the files.

Differential Revision: https://phabricator.services.mozilla.com/D114529

UltraBlame original commit: 58e693a6bde5f874543b92d1fbdcf9beb9b075f8