Bug 1706612 - Part 2: Implement the allow list for query stripping. r=dimi
This patch implements the allow listfor query stripping. The allow list
contains sites that exampt from query stripping. When loading a URI, we
will check if the basedomain of the URI is in the allow list. If it
does, we will skip the query stripping.

Depends on D115863

Differential Revision: https://phabricator.services.mozilla.com/D115864

UltraBlame original commit: d2da5189a51d3491a7e06692f820775daedcc96c